Liaison des personnages aux Users

+++ src/Controller/EtapeController.php
@@ -17,6 +17,7 @@ class EtapeController extends AbstractController
     #[Route('/', name: 'app_etape_index', methods: ['GET'])]
     public function index(EtapeRepository $etapeRepository): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_ADMIN', null, 'User tried to access a page without having ROLE_ADMIN');
         return $this->render('etape/index.html.twig', [
             'etapes' => $etapeRepository->findAll(),
         ]);
@@ -25,6 +26,7 @@ class EtapeController extends AbstractController
     #[Route('/new', name: 'app_etape_new', methods: ['GET', 'POST'])]
     public function new(Request $request, EntityManagerInterface $entityManager): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_ADMIN', null, 'User tried to access a page without having ROLE_ADMIN');
         $etape = new Etape();
         $form = $this->createForm(EtapeType::class, $etape);
         $form->handleRequest($request);
@@ -45,6 +47,7 @@ class EtapeController extends AbstractController
     #[Route('/{id}', name: 'app_etape_show', methods: ['GET'])]
     public function show(Etape $etape): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_ADMIN', null, 'User tried to access a page without having ROLE_ADMIN');
         return $this->render('etape/show.html.twig', [
             'etape' => $etape,
         ]);
@@ -53,6 +56,7 @@ class EtapeController extends AbstractController
     #[Route('/{id}/edit', name: 'app_etape_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, Etape $etape, EntityManagerInterface $entityManager): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_ADMIN', null, 'User tried to access a page without having ROLE_ADMIN');
         $form = $this->createForm(EtapeType::class, $etape);
         $form->handleRequest($request);
 
@@ -71,6 +75,7 @@ class EtapeController extends AbstractController
     #[Route('/{id}', name: 'app_etape_delete', methods: ['POST'])]
     public function delete(Request $request, Etape $etape, EntityManagerInterface $entityManager): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_ADMIN', null, 'User tried to access a page without having ROLE_ADMIN');
         if ($this->isCsrfTokenValid('delete'.$etape->getId(), $request->getPayload()->get('_token'))) {
             $entityManager->remove($etape);
             $entityManager->flush();
